+++ rasa_core/evaluate.py
@@ -9,7 +9,8 @@ import json
 import logging
 import warnings
 from builtins import str
-from typing import List, Tuple, Optional, Any, Text, Dict
+from collections import namedtuple
+from typing import List, Optional, Any, Text, Dict
 
 from sklearn.exceptions import UndefinedMetricWarning
 from tqdm import tqdm
@@ -19,15 +20,20 @@ from rasa_core import utils
 from rasa_core.agent import Agent
 from rasa_core.events import ActionExecuted, UserUttered
 from rasa_core.interpreter import NaturalLanguageInterpreter
+from rasa_core.policies import SimplePolicyEnsemble
 from rasa_core.trackers import DialogueStateTracker
 from rasa_core.training.generator import TrainingDataGenerator
 from rasa_core.utils import AvailableEndpoints, pad_list_to_size
-from rasa_nlu.evaluate import (
-    plot_confusion_matrix,
-    get_evaluation_metrics)
-from rasa_nlu.training_data.formats import MarkdownWriter, MarkdownReader
+from rasa_nlu.evaluate import plot_confusion_matrix, get_evaluation_metrics
+
 
 logger = logging.getLogger(__name__)
+
+StoryEvalution = namedtuple("StoryEvaluation",
+                            "evaluation_store "
+                            "failed_stories "
+                            "action_list "
+                            "in_training_data_fraction")
 
 
 def create_argument_parser():
@@ -140,19 +146,35 @@ class EvaluationStore(object):
                self.entity_predictions != self.entity_targets or \
                self.action_predictions != self.action_targets
 
-    def serialise_targets(self):
-        targets = self.action_targets + \
-                  self.intent_targets + \
-                  self.entity_targets
+    def serialise_targets(self,
+                          include_actions=True,
+                          include_intents=True,
+                          include_entities=False):
+        targets = []
+        if include_actions:
+            targets += self.action_targets
+        if include_intents:
+            targets += self.intent_targets
+        if include_entities:
+            targets += self.entity_targets
+
         return [json.dumps(t) if isinstance(t, dict) else t for t in targets]
 
-    def serialise_predictions(self):
-        predictions = self.action_predictions + \
-                      self.intent_predictions + \
-                      self.entity_predictions
+    def serialise_predictions(self,
+                              include_actions=True,
+                              include_intents=True,
+                              include_entities=False):
+        predictions = []
 
-        return [json.dumps(p) if isinstance(p, dict) else p
-                for p in predictions]
+        if include_actions:
+            predictions += self.action_predictions
+        if include_intents:
+            predictions += self.intent_predictions
+        if include_entities:
+            predictions += self.entity_predictions
+
+        return [json.dumps(t) if isinstance(t, dict) else t
+                for t in predictions]
 
 
 class WronglyPredictedAction(ActionExecuted):
@@ -163,15 +185,27 @@ class WronglyPredictedAction(ActionExecuted):
 
     type_name = "wrong_action"
 
-    def __init__(self, correct_action, predicted_action, timestamp=None):
-        self.correct_action = correct_action
+    def __init__(self, correct_action, predicted_action,
+                 policy, confidence, timestamp=None):
         self.predicted_action = predicted_action
         super(WronglyPredictedAction, self).__init__(correct_action,
+                                                     policy,
+                                                     confidence,
                                                      timestamp=timestamp)
 
     def as_story_string(self):
-        return "{}   <!-- predicted: {} -->".format(self.correct_action,
+        return "{}   <!-- predicted: {} -->".format(self.action_name,
                                                     self.predicted_action)
+
+
+class EndToEndUserUtterance(UserUttered):
+    """End-to-end user utterance.
+
+    Mostly used to print the full end-to-end user message in the
+    `failed_stories.md` output file."""
+
+    def as_story_string(self):
+        return super(EndToEndUserUtterance, self).as_story_string(e2e=True)
 
 
 class WronglyClassifiedUserUtterance(UserUttered):
@@ -185,42 +219,33 @@ class WronglyClassifiedUserUtterance(UserUttered):
     def __init__(self,
                  text,
                  correct_intent,
-                 predicted_intent,
-                 correct_entities=None,
-                 predicted_entities=None,
-                 timestamp=None):
-        self.text = text
-        self.correct_intent = correct_intent
+                 correct_entities,
+                 parse_data=None,
+                 timestamp=None,
+                 input_channel=None,
+                 predicted_intent=None,
+                 predicted_entities=None):
         self.predicted_intent = predicted_intent
-        self.correct_entities = correct_entities
         self.predicted_entities = predicted_entities
-        super(WronglyClassifiedUserUtterance, self).__init__(
-                text, {"name": self.correct_intent}, timestamp=timestamp)
 
-    def _deserialise_entities(self, entities):
-        if isinstance(entities, str):
-            entities = json.loads(entities)
+        intent = {"name": correct_intent}
 
-        return [e for e in entities if isinstance(e, dict)]
-
-    def _md_format_message(self, text, intent, entities):
-        message_from_md = MarkdownReader()._parse_training_example(text)
-        deserialised_entities = self._deserialise_entities(entities)
-        return MarkdownWriter()._generate_message_md(
-                {"text": message_from_md.text,
-                 "intent": intent,
-                 "entities": deserialised_entities}
-        )
+        super(WronglyClassifiedUserUtterance, self).__init__(text,
+                                                             intent,
+                                                             correct_entities,
+                                                             parse_data,
+                                                             timestamp,
+                                                             input_channel)
 
     def as_story_string(self):
-        correct_message = self._md_format_message(self.text,
-                                                  self.correct_intent,
-                                                  self.correct_entities)
-        predicted_message = self._md_format_message(self.text,
-                                                    self.predicted_intent,
-                                                    self.predicted_entities)
+        correct_message = _md_format_message(self.text,
+                                             self.intent,
+                                             self.entities)
+        predicted_message = _md_format_message(self.text,
+                                               self.predicted_intent,
+                                               self.predicted_entities)
         return ("{}: {}   <!-- predicted: {}: {} -->"
-                "").format(self.correct_intent,
+                "").format(self.intent.get("name"),
                            correct_message,
                            self.predicted_intent,
                            predicted_message)
@@ -274,16 +299,22 @@ def _collect_user_uttered_predictions(event,
     if user_uttered_eval_store.has_prediction_target_mismatch():
         partial_tracker.update(
                 WronglyClassifiedUserUtterance(
-                        event.text, intent_gold, predicted_intent,
-                        user_uttered_eval_store.entity_targets,
-                        user_uttered_eval_store.entity_predictions)
+                        event.text, intent_gold,
+                        user_uttered_eval_store.entity_predictions,
+                        event.parse_data,
+                        event.timestamp,
+                        event.input_channel,
+                        predicted_intent,
+                        user_uttered_eval_store.entity_targets)
         )
         if fail_on_prediction_errors:
             raise ValueError(
                     "NLU model predicted a wrong intent. Failed Story:"
                     " \n\n{}".format(partial_tracker.export_stories()))
     else:
-        partial_tracker.update(event)
+        end_to_end_user_utterance = EndToEndUserUtterance(
+                event.text, event.intent, event.entities)
+        partial_tracker.update(end_to_end_user_utterance)
 
     return user_uttered_eval_store
 
@@ -292,7 +323,7 @@ def _collect_action_executed_predictions(processor, partial_tracker, event,
                                          fail_on_prediction_errors):
     action_executed_eval_store = EvaluationStore()
 
-    action, _, _ = processor.predict_next_action(partial_tracker)
+    action, policy, confidence = processor.predict_next_action(partial_tracker)
 
     predicted = action.name()
     gold = event.action_name
@@ -301,7 +332,10 @@ def _collect_action_executed_predictions(processor, partial_tracker, event,
                                             action_targets=gold)
 
     if action_executed_eval_store.has_prediction_target_mismatch():
-        partial_tracker.update(WronglyPredictedAction(gold, predicted))
+        partial_tracker.update(WronglyPredictedAction(gold, predicted,
+                                                      event.policy,
+                                                      event.confidence,
+                                                      event.timestamp))
         if fail_on_prediction_errors:
             raise ValueError(
                     "Model predicted a wrong action. Failed Story: "
@@ -309,13 +343,12 @@ def _collect_action_executed_predictions(processor, partial_tracker, event,
     else:
         partial_tracker.update(event)
 
-    return action_executed_eval_store
+    return action_executed_eval_store, policy, confidence
 
 
 def _predict_tracker_actions(tracker, agent, fail_on_prediction_errors=False,
                              use_e2e=False):
     processor = agent.create_processor()
-
     tracker_eval_store = EvaluationStore()
 
     events = list(tracker.events)
@@ -324,14 +357,22 @@ def _predict_tracker_actions(tracker, agent, fail_on_prediction_errors=False,
                                                        events[:1],
                                                        agent.domain.slots)
 
+    tracker_actions = []
+
     for event in events[1:]:
         if isinstance(event, ActionExecuted):
-            action_executed_result = \
+            action_executed_result, policy, confidence = \
                 _collect_action_executed_predictions(
                         processor, partial_tracker, event,
                         fail_on_prediction_errors
                 )
             tracker_eval_store.merge_store(action_executed_result)
+            tracker_actions.append(
+                    {"action": action_executed_result.action_targets[0],
+                     "predicted": action_executed_result.action_predictions[0],
+                     "policy": policy,
+                     "confidence": confidence}
+            )
         elif use_e2e and isinstance(event, UserUttered):
             user_uttered_result = \
                 _collect_user_uttered_predictions(
@@ -341,7 +382,19 @@ def _predict_tracker_actions(tracker, agent, fail_on_prediction_errors=False,
         else:
             partial_tracker.update(event)
 
-    return tracker_eval_store, partial_tracker
+    return tracker_eval_store, partial_tracker, tracker_actions
+
+
+def _in_training_data_fraction(action_list):
+    """Given a list of action items, returns the fraction of actions
+
+    that were predicted using one of the Memoization policies."""
+    in_training_data = [
+        a["action"] for a in action_list
+        if not SimplePolicyEnsemble.is_not_memo_policy(a["policy"])
+    ]
+
+    return len(in_training_data) / len(action_list)
 
 
 def collect_story_predictions(
@@ -350,7 +403,7 @@ def collect_story_predictions(
         fail_on_prediction_errors=False,  # type: bool
         use_e2e=False  # type: bool
 ):
-    # type: (...) -> Tuple[EvaluationStore, List[DialogueStateTracker]]
+    # type: (...) -> StoryEvalution
     """Test the stories from a file, running them through the stored model."""
 
     story_eval_store = EvaluationStore()
@@ -360,12 +413,16 @@ def collect_story_predictions(
     logger.info("Evaluating {} stories\n"
                 "Progress:".format(len(completed_trackers)))
 
+    action_list = []
+
     for tracker in tqdm(completed_trackers):
-        tracker_results, predicted_tracker = \
+        tracker_results, predicted_tracker, tracker_actions = \
             _predict_tracker_actions(tracker, agent,
                                      fail_on_prediction_errors, use_e2e)
 
         story_eval_store.merge_store(tracker_results)
+
+        action_list.extend(tracker_actions)
 
         if tracker_results.has_prediction_target_mismatch():
             # there is at least one wrong prediction
@@ -378,12 +435,18 @@ def collect_story_predictions(
     report, precision, f1, accuracy = get_evaluation_metrics(
             [1] * len(completed_trackers), correct_dialogues)
 
+    in_training_data_fraction = _in_training_data_fraction(action_list)
+
     log_evaluation_table([1] * len(completed_trackers),
                          "END-TO-END" if use_e2e else "CONVERSATION",
                          report, precision, f1, accuracy,
+                         in_training_data_fraction,
                          include_report=False)
 
-    return story_eval_store, failed
+    return StoryEvalution(evaluation_store=story_eval_store,
+                          failed_stories=failed,
+                          action_list=action_list,
+                          in_training_data_fraction=in_training_data_fraction)
 
 
 def log_failed_stories(failed, failed_output):
@@ -412,43 +475,53 @@ def run_story_evaluation(resource_name, agent,
     completed_trackers = _generate_trackers(resource_name, agent,
                                             max_stories, use_e2e)
 
-    story_results, failed = collect_story_predictions(
-            completed_trackers, agent, fail_on_prediction_errors, use_e2e)
+    story_evaluation = collect_story_predictions(completed_trackers, agent,
+                                                 fail_on_prediction_errors,
+                                                 use_e2e)
+
+    evaluation_store = story_evaluation.evaluation_store
 
     with warnings.catch_warnings():
         warnings.simplefilter("ignore", UndefinedMetricWarning)
         report, precision, f1, accuracy = get_evaluation_metrics(
-                story_results.serialise_targets(),
-                story_results.serialise_predictions()
+                evaluation_store.serialise_targets(),
+                evaluation_store.serialise_predictions()
         )
 
     if out_file_plot:
-        plot_story_evaluation(story_results.action_targets,
-                              story_results.action_predictions,
-                              report, precision, f1, accuracy, out_file_plot)
+        plot_story_evaluation(evaluation_store.action_targets,
+                              evaluation_store.action_predictions,
+                              report, precision, f1, accuracy,
+                              story_evaluation.in_training_data_fraction,
+                              out_file_plot)
 
-    log_failed_stories(failed, out_file_stories)
+    log_failed_stories(story_evaluation.failed_stories, out_file_stories)
 
     return {
-        "story_evaluation": {
-            "report": report,
-            "precision": precision,
-            "f1": f1,
-            "accuracy": accuracy
-        }
+        "report": report,
+        "precision": precision,
+        "f1": f1,
+        "accuracy": accuracy,
+        "actions": story_evaluation.action_list,
+        "in_training_data_fraction":
+            story_evaluation.in_training_data_fraction,
+        "is_end_to_end_evaluation": use_e2e
     }
 
 
 def log_evaluation_table(golds, name,
                          report, precision, f1, accuracy,
+                         in_training_data_fraction,
                          include_report=True):  # pragma: no cover
     """Log the sklearn evaluation metrics."""
     logger.info("Evaluation Results on {} level:".format(name))
-    logger.info("\tCorrect:   {} / {}".format(int(len(golds) * accuracy),
-                                              len(golds)))
-    logger.info("\tF1-Score:  {:.3f}".format(f1))
-    logger.info("\tPrecision: {:.3f}".format(precision))
-    logger.info("\tAccuracy:  {:.3f}".format(accuracy))
+    logger.info("\tCorrect:          {} / {}"
+                "".format(int(len(golds) * accuracy), len(golds)))
+    logger.info("\tF1-Score:         {:.3f}".format(f1))
+    logger.info("\tPrecision:        {:.3f}".format(precision))
+    logger.info("\tAccuracy:         {:.3f}".format(accuracy))
+    logger.info("\tIn-data fraction: {:.3g}"
+                "".format(in_training_data_fraction))
 
     if include_report:
         logger.info("\tClassification report: \n{}".format(report))
@@ -456,6 +529,7 @@ def log_evaluation_table(golds, name,
 
 def plot_story_evaluation(test_y, predictions,
                           report, precision, f1, accuracy,
+                          in_training_data_fraction,
                           out_file):
     """Plot the results of story evaluation"""
     from sklearn.metrics import confusion_matrix
@@ -464,6 +538,7 @@ def plot_story_evaluation(test_y, predictions,
 
     log_evaluation_table(test_y, "ACTION",
                          report, precision, f1, accuracy,
+                         in_training_data_fraction,
                          include_report=True)
 
     cnf_matrix = confusion_matrix(test_y, predictions)