+++ rasa_core/utils.py
@@ -13,6 +13,7 @@ import os
 import re
 import sys
 import tempfile
+from builtins import input, range, str
 from hashlib import sha1
 from random import Random
 from threading import Thread
@@ -20,7 +21,6 @@ from typing import Text, Any, List, Optional, Tuple, Dict, Set
 
 import requests
 import six
-from builtins import input, range, str
 from numpy import all, array
 from requests.auth import HTTPBasicAuth
 from requests.exceptions import InvalidURL
@@ -73,6 +73,7 @@ def add_logging_option_arguments(parser):
     )
 
 
+# noinspection PyUnresolvedReferences
 def class_from_module_path(module_path):
     # type: (Text) -> Any
     """Given the module name and path of a class, tries to retrieve the class.
@@ -81,13 +82,17 @@ def class_from_module_path(module_path):
     import importlib
 
     # load the module, will raise ImportError if module cannot be loaded
+    from rasa_core.policies.keras_policy import KerasPolicy
+    from rasa_core.policies.fallback import FallbackPolicy
+    from rasa_core.policies.memoization import MemoizationPolicy
+
     if "." in module_path:
         module_name, _, class_name = module_path.rpartition('.')
         m = importlib.import_module(module_name)
         # get the class, will raise AttributeError if class cannot be found
         return getattr(m, class_name)
     else:
-        return globals()[module_path]
+        return globals().get(module_path, locals().get(module_path))
 
 
 def module_path_from_instance(inst):
@@ -466,6 +471,21 @@ def bool_arg(name, default=True):
     return request.args.get(name, str(default)).lower() == 'true'
 
 
+def float_arg(name):
+    # type: ( Text) -> float
+    """Return a passed argument cast as a float or None.
+
+    Checks the `name` parameter of the request if it contains a valid
+    float value. If not, `None` is returned."""
+    from flask import request
+
+    arg = request.args.get(name)
+    try:
+        return float(arg)
+    except TypeError:
+        return None
+
+
 def extract_args(kwargs,  # type: Dict[Text, Any]
                  keys_to_extract  # type: Set[Text]
                  ):
@@ -486,7 +506,8 @@ def extract_args(kwargs,  # type: Dict[Text, Any]
 
 
 def arguments_of(func):
-    """Return the parameters of the function `func` as a list of their names."""
+    """Return the parameters of the function `func` """
+    """as a list of their names."""
 
     try:
         # python 3.x is used
@@ -527,7 +548,7 @@ def all_subclasses(cls):
 
 def read_endpoint_config(filename, endpoint_type):
     # type: (Text, Text) -> Optional[EndpointConfig]
-    """Read an endpoint configuration file from disk and extract one config. """
+    """Read an endpoint configuration file from disk and extract one config."""
 
     if not filename:
         return None
@@ -601,27 +622,37 @@ class AvailableEndpoints(object):
                 endpoint_file, endpoint_type="action_endpoint")
         model = read_endpoint_config(
                 endpoint_file, endpoint_type="models")
+        tracker_store = read_endpoint_config(
+                endpoint_file, endpoint_type="tracker_store")
 
-        return cls(nlg, nlu, action, model)
+        return cls(nlg, nlu, action, model, tracker_store)
 
-    def __init__(self, nlg=None, nlu=None, action=None, model=None):
+    def __init__(self,
+                 nlg=None,
+                 nlu=None,
+                 action=None,
+                 model=None,
+                 tracker_store=None):
         self.model = model
         self.action = action
         self.nlu = nlu
         self.nlg = nlg
+        self.tracker_store = tracker_store
 
 
 class EndpointConfig(object):
     """Configuration for an external HTTP endpoint."""
 
     def __init__(self, url, params=None, headers=None, basic_auth=None,
-                 token=None, token_name="token"):
+                 token=None, token_name="token", **kwargs):
         self.url = url
         self.params = params if params else {}
         self.headers = headers if headers else {}
         self.basic_auth = basic_auth
         self.token = token
         self.token_name = token_name
+        self.store_type = kwargs.pop('store_type', None)
+        self.kwargs = kwargs
 
     def request(self,
                 method="post",  # type: Text
@@ -672,13 +703,7 @@ class EndpointConfig(object):
 
     @classmethod
     def from_dict(cls, data):
-        return EndpointConfig(
-                data.get("url"),
-                data.get("params"),
-                data.get("headers"),
-                data.get("basic_auth"),
-                data.get("token"),
-                data.get("token_name"))
+        return EndpointConfig(**data)
 
     def __eq__(self, other):
         if isinstance(self, type(other)):